nav-backend/requirements.txt
# ── 核心框架 ──
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# ── AgentScope（多智能体框架）──
agentscope>=0.1.0

# ── 高德地图 HTTP 客户端 ──
aiohttp>=3.9.0

# ── 数据验证 ──
pydantic>=2.0

# ── Whisper 语音识别 ──
openai-whisper>=20240930
numpy

# ── 视觉分析（qwen-vl-max 通过 OpenAI 兼容接口）──
openai>=1.0

# ── FFmpeg（Whisper 依赖，需系统安装）──
# sudo apt install ffmpeg  或  brew install ffmpeg
